# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-grade REST API for a single-gym management system in Bagerhat, Bangladesh. Two roles: Admin (gym owner) and Member (gym customer).",
        "title": "Fitness Care Bagerhat API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/members/{id}/subscriptions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/subscriptions"
                ],
                "summary": "List all subscriptions for a member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Member UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/subscriptions"
                ],
                "summary": "Assign a plan to a member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Member UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Subscription details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.assignPlanReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/admin/members/{id}/subscriptions/active": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/subscriptions"
                ],
                "summary": "Update the active subscription (price, end_date, note)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Member UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateActiveReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/plans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/plans"
                ],
                "summary": "List all plan templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/plans"
                ],
                "summary": "Create a plan template",
                "parameters": [
                    {
                        "description": "Plan details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createPlanReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/admin/plans/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin/plans"
                ],
                "summary": "Delete a plan template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/plans"
                ],
                "summary": "Update a plan template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updatePlanReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.assignPlanReq": {
            "type": "object",
            "required": [
                "plan_template_id"
            ],
            "properties": {
                "final_price": {
                    "description": "optional; defaults to plan price",
                    "type": "number"
                },
                "note": {
                    "type": "string"
                },
                "plan_template_id": {
                    "type": "string"
                },
                "start_date": {
                    "description": "YYYY-MM-DD, optional",
                    "type": "string"
                }
            }
        },
...
        "handlers.createPlanReq": {
            "type": "object",
            "required": [
                "default_price",
                "duration_days",
                "name"
            ],
            "properties": {
                "default_price": {
                    "type": "number",
                    "minimum": 0
                },
                "duration_days": {
                    "type": "integer",
                    "maximum": 3650,
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
...
        "handlers.updateActiveReq": {
            "type": "object",
            "required": [
                "end_date",
                "final_price"
            ],
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "final_price": {
                    "type": "number",
                    "minimum": 0
                },
                "note": {
                    "type": "string"
                }
            }
        },
...
        "handlers.updatePlanReq": {
            "type": "object",
            "required": [
                "default_price",
                "duration_days",
                "name"
            ],
            "properties": {
                "default_price": {
                    "type": "number",
                    "minimum": 0
                },
                "duration_days": {
                    "type": "integer",
                    "maximum": 3650,
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the Bearer prefix: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}